Adjust for illumos

+++ src/main.rs
@@ -192,9 +192,9 @@ pub enum Error {
 }
 
 // Ioctl command definitions
-const PCITOOL_IOC: u64 = (('P' as u64) << 24) | (('C' as u64) << 16) | (('T' as u64) << 8);
-const PCITOOL_DEVICE_GET_REG: u64 = PCITOOL_IOC | 1;
-const PCITOOL_DEVICE_SET_REG: u64 = PCITOOL_IOC | 2;
+const PCITOOL_IOC: i32 = (('P' as i32) << 24) | (('C' as i32) << 16) | (('T' as i32) << 8);
+const PCITOOL_DEVICE_GET_REG: i32 = PCITOOL_IOC | 1;
+const PCITOOL_DEVICE_SET_REG: i32 = PCITOOL_IOC | 2;
 
 // TODO-correctness: The register file should not matter so much, since x86 will
 // ultimately look up the device just by the BDF. But we should still do the
